+++ src/main/kotlin/gli_/Storage.kt
@@ -43,7 +43,7 @@ class Storage {
         blockCount = Vec3i(storage.blockCount)
         blockExtent = Vec3i(storage.blockExtent)
         extent = Vec3i(storage.extent)
-        data = MemoryUtil.memByteBuffer(storage.data!!.adr, storage.data!!.remaining())
+        data = MemoryUtil.memByteBuffer(storage.data!!.adr.toLong(), storage.data!!.remaining())
     }
 
     constructor(format: Format, extent: Vec3i, layers: Int, faces: Int, levels: Int) {
@@ -120,8 +120,8 @@ class Storage {
 
         val baseOffsetSrc = storageSrc.baseOffset(layerSrc, faceSrc, levelSrc)
         val baseOffsetDst = baseOffset(layerDst, faceDst, levelDst)
-        val imageSrc = storageSrc.data!!.adr + baseOffsetSrc
-        val imageDst = data!!.adr + baseOffsetDst
+        val imageSrc = storageSrc.data!!.adr.toLong() + baseOffsetSrc
+        val imageDst = data!!.adr.toLong() + baseOffsetDst
 
         for (blockIndexZ in 0 until blockCount.z)
             for (blockIndexY in 0 until blockCount.y) {
